Add MPR trashing protection for MPR-window CD_READ calls
Following many reports of crashing on real hardware and a great deal of
bewildered debugging, it turns out that there is a bug in the MPR-window
modes (2-6) of the BIOS CD_READ function that appears to have not been
previously documented.

Due to an oversight in the code that is supposed to restore the MPR used for
the read window to its original page, the MPR will only be correctly restored
if the read succeeds on the first attempt. If it fails and is retried, the
original value will be thrown away, leaving a trashed value behind when the
routine returns.

Because of this error, code that uses this version of CD_READ cannot rely on
the MPR being restored to its original value. Several calls in this
translation did exactly that, and have now been modified to ensure correct
behavior even if a read fails. (The issue went undetected because emulators
and flashcards generally simulate perfect behavior from the disc drive, which
of course isn't always the case in reality.)